Fixed  prev and Next Bug
Modified the order of the code  so NOw, The buttons responf on the first
click instead of a second click.  Also,  Because the timing was not
reseting,  It was creating an error at the end when time ran out.  Now,
LastTime resets  with every click,  This laso gves the user more time to
see the pictures while navigating